Styling and creation of the responses (success/error)
If the 00 (success) code is returned we will apply the success styling,
for everything else the error styling is applied. These are defined
within the css and are customizable via the themes sectionIf the 00
(success) code is returned we will apply the success styling, for
everything else the error styling is applied. These are defined within
the css and are customizable via the themes section.